requirements.txt
requests>=2.31.0
pdfplumber>=0.10.0
tqdm>=4.66.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
